feat: add api follow but error -2 +2, refactor id from int to long in all tables

+++ src/main/java/com/isocial/minisocialbe/controller/ProfileController.java
@@ -16,7 +16,7 @@ public class ProfileController {
     private final ProfileService profileService;
 
     @PostMapping("/{userId}")
-    public ResponseEntity<ProfileResponseDto> getProfileByUserId(@PathVariable("userId") Integer userId) {
+    public ResponseEntity<ProfileResponseDto> getProfileByUserId(@PathVariable("userId") Long userId) {
         ProfileResponseDto profile = profileService.getProfileByUserId(userId);
         return ResponseEntity.ok(profile);
     }
